refactor(pay): PG 클라이언트 인프라 패키지 이동
외부 PG 호출 인터페이스를 설정 패키지에서 인프라 패키지로 분리합니다.

Refs #435
