+++ backend/widyu-api/src/main/java/com/widyu/pay/application/PaymentService.java
@@ -11,7 +11,7 @@ import com.widyu.pay.PaymentCancel;
 import com.widyu.pay.PaymentOrder;
 import com.widyu.pay.PaymentStatus;
 import com.widyu.pay.PointChargePackage;
-import com.widyu.pay.config.PaymentClient;
+import com.widyu.pay.infrastructure.PaymentClient;
 import com.widyu.pay.dto.mapper.PaymentMapper;
 import com.widyu.pay.dto.request.CancelRequest;
 import com.widyu.pay.dto.request.PaymentApproveRequest;
