+++ backend/widyu-api/src/test/java/com/widyu/pay/integration/PaymentIntegrationTest.java
@@ -22,7 +22,7 @@ import com.widyu.pay.Payment;
 import com.widyu.pay.PaymentOrder;
 import com.widyu.pay.PaymentStatus;
 import com.widyu.pay.application.PaymentService;
-import com.widyu.pay.config.PaymentClient;
+import com.widyu.pay.infrastructure.PaymentClient;
 import com.widyu.pay.dto.request.CancelRequest;
 import com.widyu.pay.dto.request.PaymentApproveRequest;
 import com.widyu.pay.dto.request.PaymentOrderCreateRequest;
